Move equality query to own file, move equality inference to quantifiers engine.
