+++ src/theory/quantifiers_engine.h
@@ -56,6 +56,8 @@ namespace quantifiers {
   class RelevantDomain;
   class BvInverter;
   class InstPropagator;
+  class EqualityInference;
+  class EqualityQueryQuantifiersEngine;
   //modules, these are "subsolvers" of the quantifiers theory.
   class InstantiationEngine;
   class ModelEngine;
@@ -81,8 +83,6 @@ namespace inst {
   class TriggerTrie;
 }/* CVC4::theory::inst */
 
-//class EfficientEMatcher;
-class EqualityQueryQuantifiersEngine;
 
 class QuantifiersEngine {
   friend class quantifiers::InstantiationEngine;
@@ -106,7 +106,9 @@ private:
   /** instantiation notify */
   std::vector< InstantiationNotify* > d_inst_notify;
   /** equality query class */
-  EqualityQueryQuantifiersEngine* d_eq_query;
+  quantifiers::EqualityQueryQuantifiersEngine* d_eq_query;
+  /** equality inference class */
+  quantifiers::EqualityInference* d_eq_inference;
   /** for computing relevance of quantifiers */
   QuantRelevance * d_quant_rel;
   /** relevant domain */
@@ -219,7 +221,9 @@ public:
   /** get theory engine */
   TheoryEngine* getTheoryEngine() { return d_te; }
   /** get equality query */
-  EqualityQueryQuantifiersEngine* getEqualityQuery();
+  EqualityQuery* getEqualityQuery();
+  /** get the equality inference */
+  quantifiers::EqualityInference* getEqualityInference() { return d_eq_inference; }
   /** get default sat context for quantifiers engine */
   context::Context* getSatContext();
   /** get default sat context for quantifiers engine */
@@ -434,56 +438,6 @@ public:
   };/* class QuantifiersEngine::Statistics */
   Statistics d_statistics;
 };/* class QuantifiersEngine */
-
-
-
-/** equality query object using theory engine */
-class EqualityQueryQuantifiersEngine : public EqualityQuery
-{
-private:
-  /** pointer to theory engine */
-  QuantifiersEngine* d_qe;
-  /** quantifiers equality inference */
-  quantifiers::EqualityInference * d_eq_inference;
-  context::CDO< unsigned > d_eqi_counter;
-  /** internal representatives */
-  std::map< TypeNode, std::map< Node, Node > > d_int_rep;
-  /** rep score */
-  std::map< Node, int > d_rep_score;
-  /** reset count */
-  int d_reset_count;
-
-  /** processInferences : will merge equivalence classes in master equality engine, if possible */
-  bool processInferences( Theory::Effort e );
-  /** node contains */
-  Node getInstance( Node n, const std::vector< Node >& eqc, std::unordered_map<TNode, Node, TNodeHashFunction>& cache );
-  /** get score */
-  int getRepScore( Node n, Node f, int index, TypeNode v_tn );
-  /** flatten representatives */
-  void flattenRepresentatives( std::map< TypeNode, std::vector< Node > >& reps );
-public:
-  EqualityQueryQuantifiersEngine( context::Context* c, QuantifiersEngine* qe );
-  virtual ~EqualityQueryQuantifiersEngine();
-  /** reset */
-  bool reset( Theory::Effort e );
-  /** identify */
-  std::string identify() const { return "EqualityQueryQE"; }
-  /** general queries about equality */
-  bool hasTerm( Node a );
-  Node getRepresentative( Node a );
-  bool areEqual( Node a, Node b );
-  bool areDisequal( Node a, Node b );
-  eq::EqualityEngine* getEngine();
-  void getEquivalenceClass( Node a, std::vector< Node >& eqc );
-  TNode getCongruentTerm( Node f, std::vector< TNode >& args );
-  /** getInternalRepresentative gets the current best representative in the equivalence class of a, based on some criteria.
-      If cbqi is active, this will return a term in the equivalence class of "a" that does
-      not contain instantiation constants, if such a term exists.
-   */
-  Node getInternalRepresentative( Node a, Node f, int index );
-  /** get quantifiers equality inference */
-  quantifiers::EqualityInference * getEqualityInference() { return d_eq_inference; }
-}; /* EqualityQueryQuantifiersEngine */
 
 }/* CVC4::theory namespace */
 }/* CVC4 namespace */
